requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
mssql-django>=1.4.1
pyodbc>=4.0.39
python-decouple>=3.8
djangorestframework-simplejwt>=5.3.0